backend/requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
gunicorn==23.0.0

# OpenAI
openai==1.57.2

# PDF processing
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
